+++ docs/CHANGELOG.md
@@ -1,7 +1,15 @@
-Changelog
+Change log
 =========
 
-Version 0.1
------------
+Version 0.1.1
+-------------
+
+Added metrics classes (AUC, Cindex, Brier score)
+Added Kaplan Meier
+Created Sphinx documentation
+Added R benchmark comparison
+
+Version 0.1.0
+-------------
 
 Initial release of CoxPH and Weibull classes.